update grades and calculate score for lineups

+++ fanta-data-service/models.py
@@ -56,7 +56,7 @@ class Squad(SQLModel, table=True):
     owner_id: int = Field(foreign_key="user.id", ondelete="CASCADE") # Foreign Key verso User
     league_id: int = Field(foreign_key="league.id", ondelete="CASCADE") # Foreign Key verso
     name: str = Field(index=True)
-    score: int = Field(default=0)
+    score: float = Field(default=0)
 
     # comodità
     owner: User = Relationship() # user che possiede la rosa
@@ -109,7 +109,7 @@ class LineUp(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     squad_id: int = Field(foreign_key="squad.id", ondelete="CASCADE") # Foreign Key verso Squad
     matchday_id: int = Field(foreign_key="matchday.id", ondelete="RESTRICT") # Foreign Key verso MatchDay
-    score: int = Field(default=0)
+    score: float = Field(default=0)
 
     players: list[PlayerLineUpLink] = Relationship(back_populates="lineup")
     squad: Squad = Relationship(back_populates="lineups")
@@ -126,7 +126,7 @@ class LeagueUpdate(SQLModel):
 
 class SquadUpdate(SQLModel):
     name: Optional[str] = None
-    score: Optional[int] = None
+    score: Optional[float] = None
     
 class LeagueWithParticipants(SQLModel):
     id: Optional[int] = None
@@ -143,7 +143,7 @@ class SquadWithPlayers(SQLModel):
     owner_id: int
     league_id: int
     name: str
-    score: int = 0
+    score: float = 0.0
     players: list[Player] = []
 
 class PlayerInLineUp(SQLModel):
@@ -158,7 +158,7 @@ class LineUpWithPlayers(SQLModel):
 
 class LineUpUpdate(SQLModel):
     players: Optional[list[PlayerInLineUp]] = None
-    score: Optional[int] = None
+    score: Optional[float] = None
 
 class MatchDayStatusUpdate(SQLModel):
     played_so_far: int
